src/main/java/com/flyonsky/weixin/data/material/CreateOtherReceive.java
package com.flyonsky.weixin.data.material;

import com.fasterxml.jackson.annotation.JsonInclude;
import com.fasterxml.jackson.annotation.JsonProperty;

/**
 * 新增其他类型永久素材响应
 * @author Administrator
 *
 */
@JsonInclude(JsonInclude.Include.NON_NULL)
public class CreateOtherReceive extends CreateNewReceive{

	// 新增的图片素材的图片URL（仅新增图片素材时会返回该字段）
	@JsonProperty("url")
	private String url;

	public String getUrl() {
		return url;
	}

	public void setUrl(String url) {
		this.url = url;
	}
}
